Make use of SETX command to add variable to registry settings
This should allow any shell started after this script, whether it
executes this script or not, to pick up the variables.  Of course,
if you reboot, or the ssh-agent gets terminated, or what not, then
that's not going to help.  So we can't shortcut against actually
running the check when the script is run. #gklst