Attempt to guess SSH_BIN_PATH.
It's annoying for each user to have to define this for their
environment. Firstly if they're going to do that the environment is the
better place to do it. Secondly, if we can figure out which paths to
prefer we can hopefully check and "just work"(tm) saving the user the
trouble.