add -x (force exit on completion) parameter
Discovered today that GOTO :eof is actually a command, as :eof is a
reserved word that means "exit".  Therefore, the ":eof" label at the end
was never being reached!

Wanted a way to launch this from Startup, without adding a second batch
file just to exit after, so added a -X parameter that will exit from the
command interpreter completely at end (except on cases of error, in which
case it will remain open, so you can see the error)